Remove storage abstraction
